pyproject.toml
[tool.poetry]
name = "behoof"
version = "0.1.0"
description = ""
authors = ["Umbreella <draginvadim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-cors-headers = "^4.2.0"
django-filter = "^23.2"
django-leaflet = "^0.29.0"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^7.1.0" }
djangorestframework = "^3.14.0"
djangorestframework-gis = "^1.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
graphene-django = "^3.1.5"
graphene-gis = "^0.0.8"
gunicorn = "^21.2.0"
pillow = "^10.0.0"
psycopg = { extras = ["binary"], version = "^3.1.10" }
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
flake8-broken-line = "^1.0.0"
flake8-commas = "^2.1.0"
flake8-quotes = "^3.3.2"
isort = "^5.12.0"


[tool.poetry.group.test.dependencies]
coverage = "^7.3.1"
snapshottest = "^0.6.0"

[tool.isort]
multi_line_output = 5
include_trailing_comma = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
